// src/types/formik.ts
export interface Booking {
  title: string;
  name: string;
  surname: string;
  bday: string;
  month: string;
  year: string;
  email: string;
  phone: string;
  address: string;
  city: string;
  state: string;
  postal: string;
  country: string;
  departure: string;
  cityFrom: string;
  cityTo: string | undefined;
  seat: string;
}